requirements.txt
numpy~=1.26.0
pandas~=2.1.1
plotly~=5.9.0
scikit-learn~=1.3.0
pillow~=9.4.0